Backend now keeps track of when the currently playing song started

+++ src/main/java/tubalubback/models/MusicSyncInfo.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Data
 public class MusicSyncInfo {
-    private long time = 0;
+    private int time = 0;
     private List<String> songQ = new ArrayList<>();
     private List<String> history = new ArrayList<>();
 }
